Moved the function debug decorator from scons_args.py to util.py.
